fix: address 4 known issues discovered during the m1.5 loop run
Fixes four of the six pre-existing issues recorded in specs/m1.5/README.md's
known-issues section (the two deferred -- SaveOutbox's concurrent-write
race and the AgentKind/ProfileKind Nostr kind collision -- need a locking
design and a protocol-compatibility decision respectively, out of scope
for this cleanup pass):

1. internal/group/db.go: generateGroupID appends a 4-byte crypto/rand
   suffix. A nanosecond timestamp alone collided under concurrent
   CreateGroup calls from the same creator (PR #18 review: 2/9 runs).
   New TestGenerateGroupID_ConcurrentSameCreatorNoCollisions (30
   concurrent creates) fails reliably against the old implementation and
   passes reliably against the fix.

2. internal/profile: --price-min/--price-max are now cli.FloatFlag
   (were cli.IntFlag), matching --rating-min and RateEntry.Price's own
   float64 type -- a fractional boundary like --price-max 99.99 can now
   be expressed on the CLI at all, not just compared correctly
   internally (that part was already fixed in PR #20).

3. internal/messaging: event.Sign()'s previously-ignored error is now
   checked (agent.go:154) -- a signing failure aborts before publish
   instead of silently proceeding with a zero-value event.ID. Separately,
   AddToOutbox now stores OutboxEntry.ID hex-encoded rather than raw
   event.ID bytes, which used to get irreversibly mangled by
   json.Marshal's UTF-8 replacement the first time it hit disk (9 of 13
   real historical entries were already corrupted this way). The one
   other call site that constructs this lookup key directly from
   event.ID (agent.go's post-publish outbox cleanup) is updated to match;
   `storage outbox list`/`retry --id` keep working against both old
   (raw-bytes) and new (hex) entries without needing a data migration.

4. scripts/deploy-relay.sh: check_port rejects all-digit strings over 5
   characters before the numeric comparison, closing an integer-overflow
   bypass (a 20+ digit port string previously errored the comparison
   silently and fell through to reporting the port "free"). check_relay_repo
   now bounds `git ls-remote` with a new run_with_timeout helper (prefers
   timeout/gtimeout, falls back to a portable bash polling loop) instead
   of relying on the OS's own TCP connect timeout (60-120s+) against a
   blackholed host.

+++ internal/profile/commands.go
@@ -388,11 +388,11 @@ var profileDiscoverCmd = &cli.Command{
 			Name:  "capability",
 			Usage: "Only show profiles with this capability (structured-mode profiles only)",
 		},
-		&cli.IntFlag{
+		&cli.FloatFlag{
 			Name:  "price-min",
 			Usage: "Only show profiles with a rate at or above this price (structured-mode profiles only)",
 		},
-		&cli.IntFlag{
+		&cli.FloatFlag{
 			Name:  "price-max",
 			Usage: "Only show profiles with a rate at or below this price (structured-mode profiles only)",
 		},
@@ -415,11 +415,11 @@ var profileDiscoverCmd = &cli.Command{
 		var discoverFilter DiscoverFilter
 		discoverFilter.Capability = c.String("capability")
 		if c.IsSet("price-min") {
-			v := int(c.Int("price-min"))
+			v := c.Float("price-min")
 			discoverFilter.PriceMin = &v
 		}
 		if c.IsSet("price-max") {
-			v := int(c.Int("price-max"))
+			v := c.Float("price-max")
 			discoverFilter.PriceMax = &v
 		}
 		if c.IsSet("rating-min") {
